requirements.txt
openai-whisper
instaloader
requests
python-docx
langdetect
transformers
torch
PyYAML
reportlab
